+++ examples/instruments/synth_instrument.h
@@ -30,7 +30,7 @@ class SynthInstrument : public GenericInstrument {
   void Process(float* output, int num_channels,
                int num_frames) noexcept override;
   void SetData(void* /*data*/) noexcept override {}
-  void SetParameter(int index, float value) noexcept override;
+  void SetParameter(int index, double value) noexcept override;
 
   /// Returns instrument definition.
   static BarelyInstrumentDefinition GetDefinition() noexcept;
